dracut: perform special actions only if installing
Not sure `rm -f /etc/initrd-release` is needed at all anymore
but we definitely don't want to `dmesg -n 1` on every boot
considering we are going to embed the installer in the final
FCOS initrd.

+++ dracut-module/parse-coreos.sh
@@ -36,8 +36,9 @@ then
 	echo 1 > /tmp/skip_reboot
 fi
 
-# Suppress initrd-switch-root.service from starting
-rm -f /etc/initrd-release
-
-# Suppress most console messages for the installer to run without interference
-dmesg -n 1
+if $(getarg coreos.inst=); then
+    # Suppress initrd-switch-root.service from starting
+    rm -f /etc/initrd-release
+    # Suppress most console messages for the installer to run without interference
+    dmesg -n 1
+fi
